+++ packages/terra-date-picker/tests/wdio/date-picker-spec.js
@@ -1,3 +1,4 @@
+// Color contrast will be resolved in https://github.com/cerner/terra-framework/issues/744
 const ignoredA11y = {
   'color-contrast': { enabled: false },
   label: { enabled: false },